Added work_tab.html
I created another page to hold the tab information and relized just how
the javascript will work separately for each and every .html file it
uses, which is a bummer but also really cool. I also made the button
function to only get called when you press it inside the new page!